% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_AW_sleep_results.R
\name{analyze_AW_sleep_results}
\alias{analyze_AW_sleep_results}
\title{Analyze AW Sleep Data}
\usage{
analyze_AW_sleep_results(data, cutoff_hrs_for_mutiple_sleep_periods)
}
\description{
\code{analyze_AW_sleep_results} runs the standard analyisis on results from the 
read in AW function.
}
\details{
Analizes sleep data, classifies naps vs. regular sleep periods, and 
calls friday night and saturday night sleep periods "WEEKEND SLEEP"
}
