% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_AW_sleep_results_basic.R
\name{analyze_AW_sleep_results_basic}
\alias{analyze_AW_sleep_results_basic}
\title{Analyze AW Sleep Data Basic}
\usage{
analyze_AW_sleep_results_basic(data, cutoff_hrs_for_mutiple_sleep_periods)
}
\description{
\code{analyze_AW_sleep_results_basic} runs the standard analyisis on results from the 
read in AW function.
}
\details{
Analizes sleep data
}
