% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_vars.R
\name{arrange.vars}
\alias{arrange.vars}
\title{Arrange variables in a data frame.}
\usage{
arrange.vars(data, vars)
}
\description{
\code{arrange.vars} returns data frame with variables in new location.
}
\details{
Expand here
}
\examples{
arrange.vars(table, c("Var4"=1, "Var3"=2, "Var1"=4))

}
