% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_sd_fxn.R
\name{avg_sd_fxn}
\alias{avg_sd_fxn}
\title{Function to be used by summaryBy}
\usage{
avg_sd_fxn(x, n.digits = 1, include.n = FALSE)
}
\description{
Convenience function for the \code{summaryBy} function from
doBy library.
}
\details{
Convenience function for the \code{summaryBy} function from
doBy library.
}
\examples{

#Summary without sample size:
doBy::summaryBy(age ~ group, data = sleep,
                FUN = avg_sd_fxn,
                keep.names = TRUE)

#To include sample size:
doBy::summaryBy(age ~ group, data = sleep,
                FUN = function(x) {avg_sd_fxn(x, include.n = TRUE)},
                keep.names = TRUE)

}
