% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drag_down.R
\name{drag_down}
\alias{drag_down}
\title{Drag down ID variables when only the 
first row of subjects data is identified.}
\usage{
drag_down(data, name)
}
\description{
\code{drag_down} returns data frame with filled in ID variable.
}
\details{
Expand here
}
\examples{
drag_down(data, "variable")

}
