% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_actiwatch_exports.R
\name{read_actiwatch_exports}
\alias{read_actiwatch_exports}
\title{Read in actiwatch sleep data}
\usage{
read_actiwatch_exports(path)
}
\description{
\code{read_actiwatch_exports} Reads in to R all of the data from a batch
export of actiwatch sleep data.
}
\details{
Path should be to a folder where all of
the exported actiwatch .csv files are located.
}
